Port TACZ-Refabricated from MC 1.21.1 to 1.21.11
- Update all dependencies (Fabric API, Loom, JEI, REI, CCA, Cloth Config,
  ModMenu, Forge Config API Port, Player Animation Library, etc.)
- Refactor Block/Item registration to use Properties.setId() factory pattern
  required by 1.21.11
- Add mochafloats dependency for Player Animation Library compatibility
- Fix ReloadableResourcesMixin for new ReloadableServerRegistries.LoadResult
- Apply lenient JSON parsing (Strictness.LENIENT) across all resource loaders
  to handle GSON 2.11 strict mode default
- Update recipe ingredient format from object to string syntax
  ({"tag":"x"} -> "#x", {"item":"x"} -> "x")
- Update entity type tags for per-variant boat types (1.21.2+ split)
- Fix all compilation errors from MC API changes across 752 Java files
- Comment out compat modules without 1.21.11 builds (Iris, Sodium,
  shoulder surfing, controllable, carry-on, tweakeroo, etc.)

+++ src/main/java/cn/sh1rocu/tacz/mixin/accessor/GameRendererAccessor.java
@@ -9,7 +9,7 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 @Mixin(GameRenderer.class)
 public interface GameRendererAccessor {
     @Invoker("getFov")
-    double tacz$getFov(Camera camera, float tickDelta, boolean useFovSetting);
+    float tacz$getFov(Camera camera, float tickDelta, boolean useFovSetting);
 
     @Invoker("bobHurt")
     void tacz$bobHurt(PoseStack poseStack, float tickDelta);
